First working version
First working version, a lot of static coding. Could do with some flexibility but it's working pretty well. Obviously none of this is helpful with no look at the sheet the data is writing to, so I'll provide an excel shit mimicking the schema of that sheet. I will also add a whole lot more in code documentation